fix docker package
Go modules had pulled down the latest semvar tag, which
was actually super old, had to specify the commit hash
of github.com/moby/moby to get the actual up-to-date
module.

Fixed ipfs package dependencies as they are now go module
compatible and so indirect deps don't need to be listed in
this packages go.mod file.

Fixed tests so that they didn't attempt to pull a private
image, as well as not running TestRemoveAllImages by default,
as it does as it says...

Fixed util to take the official docker images into account,
as their url is docker.io/library/<container> but the container
name is just <container>, instead of library/<container>.